refactor: restructure project layout for better modularity
