+++ src/test/kotlin/lotto/util/InputValidatorTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.util
 
-import lotto.util.InputValidator
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
